PLT-1157 - add support for Solana SPL tokens (#478)
* PLT-1157 - add support for Solana SPL tokens

* Release 2.0.1-alpha.251

+++ packages/blockchain/polygon/src/lib/polygon.sdk.spec.ts
@@ -53,7 +53,7 @@ describe('TatumPolygonSDK', () => {
         const gasPrice = await sdk.getGasPriceInWei()
 
         expect(gasPrice).toBeDefined()
-        expect(parseInt(gasPrice)).toBeGreaterThan(0)
+        expect(parseFloat(gasPrice)).toBeGreaterThan(0)
       })
     })
   })
